Authorization REST API finish!,

+++ udemy/rest_api_advance/server/controllers/feed.js
@@ -1,6 +1,7 @@
 const { validationResult } = require("express-validator");
 
 const Post = require("../model/post");
+const User = require("../model/user");
 
 const fs = require("fs");
 const path = require("path");
@@ -54,20 +55,31 @@ exports.createPost = (req, res, next) => {
   }
 
   const imageUrl = req.file.path;
+  let creator;
 
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "AS MOHAMED FAHEEM ANVER" }
+    creator: req.userId
   });
 
   post
     .save()
     .then(result => {
+      oResult = result;
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post create successfully!",
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -127,6 +139,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (!imageUrl) {
         imageUrl = post.imageUrl;
       }
@@ -160,10 +178,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 400;
         throw error;
       }
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(_ => {
       res.status(200).json({ message: "Deleted!" });
     })
     .catch(err => {
